Update README.md
I have been studying database design online, realizing I had the entire concept totally wrong in my terrible data structure schemes I have attempted in the past.  Now i am learning about serverless systems and new cloud-based content and CDN's and so on, to the extent I understand even less now.  The data structure is the great challenge in this project, and it must perform on par with anything recently released, so the technology can grow and adapt to the unknown things we won't anticipate or that will change some major element in a profound way requiring major revisions, but that's tech, I guess.